cleaner web types, cleaned up imports
